requirements.txt
Flask==2.3.2
openai>=1.0.0
python-dotenv>=0.19.0
gunicorn==21.2.0
flask_cors>=4.0.0